+++ uce.portal/uce.analysis/src/main/java/org/texttechnologylab/History.java
@@ -1,5 +1,8 @@
 package org.texttechnologylab;
 
+import org.texttechnologylab.modules.DUUIInformation;
+import org.texttechnologylab.modules.ModelGroup;
+
 import java.util.HashMap;
 import java.util.List;
