+++ uce.portal/uce.analysis/src/main/java/org/texttechnologylab/TypeClasses/EmotionInput.java
@@ -1,4 +1,4 @@
-package org.texttechnologylab;
+package org.texttechnologylab.TypeClasses;
 
 public class EmotionInput {
