+++ uce.portal/uce.analysis/src/main/java/org/texttechnologylab/TypeClasses/ReadabilityClass.java
@@ -1,4 +1,6 @@
-package org.texttechnologylab;
+package org.texttechnologylab.TypeClasses;
+
+import org.texttechnologylab.modules.ModelInfo;
 
 public class ReadabilityClass {
     private double FleschKincaid;
