+++ uce.portal/uce.analysis/src/main/java/org/texttechnologylab/TypeClasses/TextClass.java
@@ -1,4 +1,6 @@
-package org.texttechnologylab;
+package org.texttechnologylab.TypeClasses;
+
+import org.texttechnologylab.modules.ModelInfo;
 
 import java.util.ArrayList;
 import java.util.HashMap;
