+++ uce.portal/uce.analysis/src/main/java/org/texttechnologylab/modules/ModelInfo.java
@@ -1,4 +1,4 @@
-package org.texttechnologylab;
+package org.texttechnologylab.modules;
