+++ uce.portal/uce.analysis/src/main/java/org/texttechnologylab/modules/ModelResources.java
@@ -1,4 +1,4 @@
-package org.texttechnologylab;
+package org.texttechnologylab.modules;
 
 import com.google.gson.Gson;
 import com.google.gson.internal.LinkedTreeMap;
@@ -10,8 +10,6 @@ import java.io.InputStreamReader;
 import java.nio.charset.StandardCharsets;
 import java.util.*;
 import java.util.stream.Collectors;
-import org.texttechnologylab.ModelInfo;
-import org.texttechnologylab.ModelGroup;
 
 public final class ModelResources {
