+++ uce.portal/uce.analysis/src/main/resources/Emotions.json
@@ -1,0 +1,16 @@
+[
+   "love",
+   "joy",
+    "surprise",
+    "anger",
+    "sadness",
+    "fear",
+    "disgust",
+    "others",
+    "anticipation",
+    "Anger",
+    "Apprehension",
+    "Happiness",
+    "Confusion",
+    "Sadness"
+]
