+++ uce.portal/uce.web/pom.xml
@@ -85,6 +85,12 @@
             <version>1.0-SNAPSHOT</version>
             <scope>compile</scope>
         </dependency>
+        <dependency>
+            <groupId>org.texttechnologylab</groupId>
+            <artifactId>uce.analysis</artifactId>
+            <version>1.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
 
         <!-- logging -->
         <dependency>
