+++ uce.portal/uce.web/src/main/java/org/texttechnologylab/LanguageResources.java
@@ -9,10 +9,7 @@ import spark.Request;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
-import java.net.URISyntaxException;
 import java.nio.charset.StandardCharsets;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.stream.Collectors;
 
 public final class LanguageResources {
